requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.1
paho-mqtt==1.6.1
pyserial==3.5
python-dotenv==1.0.0
websockets==12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6 